+++ app/src/main/java/com/hyeeyoung/wishboard/designsystem/component/button/CartButton.kt
@@ -1,16 +1,12 @@
-package com.hyeeyoung.wishboard.designsystem.component
+package com.hyeeyoung.wishboard.designsystem.component.button
 
-import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -18,7 +14,6 @@ import com.hyeeyoung.wishboard.R
 import com.hyeeyoung.wishboard.designsystem.style.WishBoardTheme
 import com.hyeeyoung.wishboard.util.extension.noRippleClickable
 
-// 패키지 이동
 @Composable
 fun CartButton(isInCart: Boolean, changeCartState: (Boolean) -> Unit) {
     val btnColor = if (isInCart) WishBoardTheme.colors.green500 else WishBoardTheme.colors.white
@@ -33,22 +28,6 @@ fun CartButton(isInCart: Boolean, changeCartState: (Boolean) -> Unit) {
             color = WishBoardTheme.colors.gray700,
         )
     }
-}
-
-@Composable
-fun WishBoardIconButton(
-    modifier: Modifier = Modifier, // TODO 디폴트 값 수정 고려 필요
-    @DrawableRes iconRes: Int,
-    onClick: () -> Unit,
-    contentDescription: String? = null,
-) {
-    Icon(
-        modifier = modifier
-            .noRippleClickable { onClick() }
-            .padding(8.dp),
-        painter = painterResource(id = iconRes),
-        contentDescription = contentDescription,
-    )
 }
 
 @Composable
